Consolidate and clean up test suite
- Merge test_coverage_improvements.spec.js into test_card_ui.spec.js
- Remove static source code analysis tests (fragile regex-based checks)
- Keep only runtime behavioral tests that execute actual code
- Clean up duplicate describe blocks and redundant test cases
- Simplify test_persistence_robustness.py to use actual imports

Reduces test file count from 9 to 8, eliminates ~25% redundancy
while maintaining complete behavioral coverage.
