+++ tests/test_persistence_robustness.py
@@ -1,967 +1,23 @@
 """Tests for persistence robustness features.
 
-These tests define the expected behavior for:
-1. Retry failed saves with exponential backoff
-2. Validate stored data on load
-
-These tests are written TDD-style - they will FAIL until the features are implemented.
-
-Note: Due to Python 3.12+ syntax in the source module, we recreate the expected
-function signatures locally and test against those. The tests validate that:
-1. The source code contains the expected retry/validation logic
-2. The recreated functions behave as expected (acting as specification)
+These tests verify:
+1. Deleted automation cleanup on load
+2. Data validation during load
+3. Graceful handling of corrupted storage
 """
 
 from __future__ import annotations
 
-import asyncio
-from dataclasses import dataclass, field
 from datetime import datetime, timedelta, timezone
-from pathlib import Path
-from typing import Any
 from unittest.mock import AsyncMock, MagicMock, patch
-import re
 
 import pytest
 
 UTC = timezone.utc
 
 
-# =============================================================================
-# Recreate data models for testing (avoids Python 3.12+ syntax issues)
-# =============================================================================
-
-
-@dataclass
-class PausedAutomation:
-    """Represent a snoozed automation."""
-
-    entity_id: str
-    friendly_name: str
-    resume_at: datetime
-    paused_at: datetime
-    days: int = 0
-    hours: int = 0
-    minutes: int = 0
-
-    def to_dict(self) -> dict[str, Any]:
-        """Convert to dictionary for storage/attributes."""
-        return {
-            "friendly_name": self.friendly_name,
-            "resume_at": self.resume_at.isoformat(),
-            "paused_at": self.paused_at.isoformat(),
-            "days": self.days,
-            "hours": self.hours,
-            "minutes": self.minutes,
-        }
-
-    @classmethod
-    def from_dict(cls, entity_id: str, data: dict[str, Any]) -> PausedAutomation:
-        """Create from dictionary."""
-        return cls(
-            entity_id=entity_id,
-            friendly_name=data.get("friendly_name", entity_id),
-            resume_at=datetime.fromisoformat(data["resume_at"]),
-            paused_at=datetime.fromisoformat(data["paused_at"]),
-            days=data.get("days", 0),
-            hours=data.get("hours", 0),
-            minutes=data.get("minutes", 0),
-        )
-
-
-@dataclass
-class ScheduledSnooze:
-    """Represent a scheduled future snooze."""
-
-    entity_id: str
-    friendly_name: str
-    disable_at: datetime
-    resume_at: datetime
-
-    def to_dict(self) -> dict[str, Any]:
-        """Convert to dictionary for storage/attributes."""
-        return {
-            "friendly_name": self.friendly_name,
-            "disable_at": self.disable_at.isoformat(),
-            "resume_at": self.resume_at.isoformat(),
-        }
-
-    @classmethod
-    def from_dict(cls, entity_id: str, data: dict[str, Any]) -> ScheduledSnooze:
-        """Create from dictionary."""
-        return cls(
-            entity_id=entity_id,
-            friendly_name=data.get("friendly_name", entity_id),
-            disable_at=datetime.fromisoformat(data["disable_at"]),
-            resume_at=datetime.fromisoformat(data["resume_at"]),
-        )
-
-
-@dataclass
-class AutomationPauseData:
-    """Runtime data for AutoSnooze."""
-
-    paused: dict[str, PausedAutomation] = field(default_factory=dict)
-    scheduled: dict[str, ScheduledSnooze] = field(default_factory=dict)
-    timers: dict[str, Any] = field(default_factory=dict)
-    scheduled_timers: dict[str, Any] = field(default_factory=dict)
-    listeners: list[Any] = field(default_factory=list)
-    store: Any = None
-
-    def notify(self) -> None:
-        """Notify all listeners of state change."""
-        for listener in self.listeners:
-            listener()
-
-
-# =============================================================================
-# Helper to read source code
-# =============================================================================
-
-
-def get_source_code() -> str:
-    """Read the source code of coordinator.py (where persistence logic lives)."""
-    coordinator_path = Path(__file__).parent.parent / "custom_components" / "autosnooze" / "coordinator.py"
-    return coordinator_path.read_text()
-
-
-def get_const_source_code() -> str:
-    """Read the source code of const.py (where constants are defined)."""
-    const_path = Path(__file__).parent.parent / "custom_components" / "autosnooze" / "const.py"
-    return const_path.read_text()
-
-
-def get_async_save_function(source: str) -> str:
-    """Extract the async_save function from source."""
-    match = re.search(r"async def async_save\([^)]+\)[^:]*:.*?(?=\n(?:async )?def |\nclass |\Z)", source, re.DOTALL)
-    return match.group(0) if match else ""
-
-
-def get_async_load_stored_function(source: str) -> str:
-    """Extract the async_load_stored function from source."""
-    match = re.search(
-        r"async def async_load_stored\([^)]+\)[^:]*:.*?(?=\n(?:async )?def |\nclass |\Z)", source, re.DOTALL
-    )
-    return match.group(0) if match else ""
-
-
-# =============================================================================
-# CURRENT IMPLEMENTATION (for comparison - this is what exists now)
-# =============================================================================
-
-
-async def _async_save_current(data: AutomationPauseData) -> None:
-    """Current implementation - no retry, no return value."""
-    if data.store is None:
-        return
-
-    try:
-        await data.store.async_save(
-            {
-                "paused": {k: v.to_dict() for k, v in data.paused.items()},
-                "scheduled": {k: v.to_dict() for k, v in data.scheduled.items()},
-            }
-        )
-    except Exception:
-        pass  # Silently ignores errors
-
-
-# =============================================================================
-# EXPECTED IMPLEMENTATION (what we want after feature is implemented)
-# =============================================================================
-
-# Transient errors that should trigger retry
-TRANSIENT_ERRORS = (IOError, OSError)
-MAX_RETRIES = 3
-BACKOFF_DELAYS = [0.1, 0.2, 0.4]  # Exponential backoff
-
-
-async def _async_save_expected(
-    data: AutomationPauseData,
-    _logger: Any = None,
-) -> bool:
-    """Expected implementation with retry logic.
-
-    Returns:
-        True if save succeeded, False if all retries exhausted.
-    """
-    if data.store is None:
-        return True  # Nothing to save, consider success
-
-    save_data = {
-        "paused": {k: v.to_dict() for k, v in data.paused.items()},
-        "scheduled": {k: v.to_dict() for k, v in data.scheduled.items()},
-    }
-
-    last_error = None
-    for attempt in range(MAX_RETRIES + 1):  # Initial + retries
-        try:
-            await data.store.async_save(save_data)
-            return True
-        except TRANSIENT_ERRORS as err:
-            last_error = err
-            if attempt < MAX_RETRIES:
-                if _logger:
-                    _logger.warning(
-                        "Save attempt %d failed, retrying in %.1fs: %s",
-                        attempt + 1,
-                        BACKOFF_DELAYS[attempt],
-                        err,
-                    )
-                await asyncio.sleep(BACKOFF_DELAYS[attempt])
-        except Exception as err:
-            # Non-transient error - don't retry
-            if _logger:
-                _logger.error("Failed to save data: %s", err)
-            return False
-
-    # All retries exhausted
-    if _logger:
-        _logger.error(
-            "Failed to save data after %d attempts: %s",
-            MAX_RETRIES + 1,
-            last_error,
-        )
-    return False
-
-
-def _validate_stored_entry(
-    entity_id: str,
-    data: dict[str, Any],
-    entry_type: str,
-    _logger: Any = None,
-) -> bool:
-    """Validate a single stored entry.
-
-    Args:
-        entity_id: The entity ID (key in storage)
-        data: The entry data
-        entry_type: "paused" or "scheduled"
-        _logger: Optional logger for warnings
-
-    Returns:
-        True if valid, False if invalid.
-    """
-    # Validate entity ID format
-    if not entity_id.startswith("automation."):
-        if _logger:
-            _logger.warning(
-                "Invalid entity_id %s: not an automation entity",
-                entity_id,
-            )
-        return False
-
-    # Validate data is a dict
-    if not isinstance(data, dict):
-        if _logger:
-            _logger.warning(
-                "Invalid data for %s: expected dict, got %s",
-                entity_id,
-                type(data).__name__,
-            )
-        return False
-
-    # Validate required fields exist
-    if entry_type == "paused":
-        required = ["resume_at", "paused_at"]
-    else:  # scheduled
-        required = ["disable_at", "resume_at"]
-
-    for field in required:
-        if field not in data:
-            if _logger:
-                _logger.warning(
-                    "Invalid data for %s: missing required field '%s'",
-                    entity_id,
-                    field,
-                )
-            return False
-
-    # Validate datetime fields
-    for field in required:
-        try:
-            datetime.fromisoformat(data[field])
-        except (ValueError, TypeError) as err:
-            if _logger:
-                _logger.warning(
-                    "Invalid data for %s: invalid datetime in '%s': %s",
-                    entity_id,
-                    field,
-                    err,
-                )
-            return False
-
-    # Validate numeric fields are non-negative (for paused entries)
-    if entry_type == "paused":
-        for field in ["days", "hours", "minutes"]:
-            value = data.get(field, 0)
-            if not isinstance(value, (int, float)) or value < 0:
-                if _logger:
-                    _logger.warning(
-                        "Invalid data for %s: %s must be non-negative, got %s",
-                        entity_id,
-                        field,
-                        value,
-                    )
-                return False
-
-    # Validate scheduled entry has disable_at < resume_at
-    if entry_type == "scheduled":
-        disable_at = datetime.fromisoformat(data["disable_at"])
-        resume_at = datetime.fromisoformat(data["resume_at"])
-        if resume_at <= disable_at:
-            if _logger:
-                _logger.warning(
-                    "Invalid scheduled snooze for %s: resume_at must be after disable_at",
-                    entity_id,
-                )
-            return False
-
-    return True
-
-
-def _validate_stored_data(
-    stored: Any,
-    _logger: Any = None,
-) -> dict[str, Any]:
-    """Validate and sanitize stored data.
-
-    Returns validated data with invalid entries removed.
-    """
-    # Handle completely invalid storage
-    if not isinstance(stored, dict):
-        if _logger:
-            _logger.error(
-                "Corrupted storage: expected dict, got %s",
-                type(stored).__name__,
-            )
-        return {"paused": {}, "scheduled": {}}
-
-    result = {"paused": {}, "scheduled": {}}
-    invalid_count = 0
-
-    # Validate paused entries
-    paused = stored.get("paused", {})
-    if isinstance(paused, dict):
-        for entity_id, entry_data in paused.items():
-            if _validate_stored_entry(entity_id, entry_data, "paused", _logger):
-                result["paused"][entity_id] = entry_data
-            else:
-                invalid_count += 1
-    else:
-        if _logger:
-            _logger.warning(
-                "Invalid 'paused' schema: expected dict, got %s",
-                type(paused).__name__,
-            )
-
-    # Validate scheduled entries
-    scheduled = stored.get("scheduled", {})
-    if isinstance(scheduled, dict):
-        for entity_id, entry_data in scheduled.items():
-            if _validate_stored_entry(entity_id, entry_data, "scheduled", _logger):
-                result["scheduled"][entity_id] = entry_data
-            else:
-                invalid_count += 1
-    else:
-        if _logger:
-            _logger.warning(
-                "Invalid 'scheduled' schema: expected dict, got %s",
-                type(scheduled).__name__,
-            )
-
-    if invalid_count > 0 and _logger:
-        _logger.info(
-            "Skipped %d invalid entries during storage load",
-            invalid_count,
-        )
-
-    return result
-
-
-# =============================================================================
-# RETRY FAILED SAVES TESTS - Source Code Verification
-# =============================================================================
-
-
-class TestRetrySaveSourceCode:
-    """Tests that verify the source code has retry logic implemented.
-
-    These tests check the actual source code for the expected patterns,
-    which will fail until the feature is implemented.
-    """
-
-    def test_async_save_has_retry_loop(self) -> None:
-        """Test that _async_save has a retry loop."""
-        source = get_source_code()
-        func = get_async_save_function(source)
-
-        # Should have a loop for retries
-        has_retry_pattern = (
-            "for attempt" in func or "for _ in range" in func or "retry" in func.lower() or "while" in func
-        )
-        assert has_retry_pattern, (
-            "FEATURE NOT IMPLEMENTED: _async_save should have retry logic. "
-            "Current implementation just catches and ignores exceptions."
-        )
-
-    def test_async_save_has_exponential_backoff(self) -> None:
-        """Test that _async_save uses exponential backoff delays."""
-        source = get_source_code()
-        func = get_async_save_function(source)
-
-        # Should use asyncio.sleep for backoff
-        has_sleep = "asyncio.sleep" in func or "await sleep" in func
-        assert has_sleep, "FEATURE NOT IMPLEMENTED: _async_save should use asyncio.sleep for backoff delays"
-
-    def test_async_save_returns_bool(self) -> None:
-        """Test that _async_save returns a boolean success status."""
-        source = get_source_code()
-        func = get_async_save_function(source)
-
-        # Should have return True and return False
-        has_return_true = "return True" in func
-        has_return_false = "return False" in func
-
-        assert has_return_true and has_return_false, (
-            "FEATURE NOT IMPLEMENTED: _async_save should return True on success, False on failure. "
-            "Current implementation returns None."
-        )
-
-    def test_async_save_catches_transient_errors(self) -> None:
-        """Test that async_save specifically catches IOError/OSError."""
-        source = get_source_code()
-        const_source = get_const_source_code()
-        func = get_async_save_function(source)
-
-        # Can use direct exception names or a TRANSIENT_ERRORS constant
-        catches_transient = "IOError" in func or "OSError" in func or "TRANSIENT_ERRORS" in func
-        # Also check that TRANSIENT_ERRORS is defined with IOError/OSError in const.py
-        if "TRANSIENT_ERRORS" in func:
-            has_transient_def = "IOError" in const_source and "OSError" in const_source
-            catches_transient = catches_transient and has_transient_def
-
-        assert catches_transient, "FEATURE NOT IMPLEMENTED: async_save should catch IOError/OSError for retry"
-
-    def test_async_save_logs_retry_attempts(self) -> None:
-        """Test that _async_save logs retry attempts."""
-        source = get_source_code()
-        func = get_async_save_function(source)
-
-        # Should log warnings on retry
-        has_warning_log = "_LOGGER.warning" in func
-        has_retry_mention = "retry" in func.lower() or "attempt" in func.lower()
-
-        assert has_warning_log and has_retry_mention, (
-            "FEATURE NOT IMPLEMENTED: _async_save should log warnings on retry attempts"
-        )
-
-
-# =============================================================================
-# RETRY FAILED SAVES TESTS - Behavior Specification
-# =============================================================================
-
-
-class TestRetrySaveBehavior:
-    """Tests that specify the expected retry behavior.
-
-    These tests use the expected implementation to verify the behavior is correct.
-    They serve as a specification for when the feature is implemented.
-    """
-
-    @pytest.fixture
-    def mock_store(self) -> MagicMock:
-        """Create mock store."""
-        store = MagicMock()
-        store.async_save = AsyncMock()
-        return store
-
-    @pytest.fixture
-    def data_with_automation(self, mock_store: MagicMock) -> AutomationPauseData:
-        """Create data with one paused automation."""
-        data = AutomationPauseData(store=mock_store)
-        now = datetime.now(UTC)
-        data.paused["automation.test"] = PausedAutomation(
-            entity_id="automation.test",
-            friendly_name="Test Automation",
-            resume_at=now + timedelta(hours=1),
-            paused_at=now,
-        )
-        return data
-
-    @pytest.mark.asyncio
-    async def test_successful_save_returns_true(
-        self, data_with_automation: AutomationPauseData, mock_store: MagicMock
-    ) -> None:
-        """Test successful save returns True."""
-        result = await _async_save_expected(data_with_automation)
-        assert result is True
-
-    @pytest.mark.asyncio
-    async def test_retries_on_transient_failure(
-        self, data_with_automation: AutomationPauseData, mock_store: MagicMock
-    ) -> None:
-        """Test that save retries on IOError."""
-        mock_store.async_save = AsyncMock(side_effect=[IOError("Disk full"), IOError("Disk full"), None])
-
-        with patch("asyncio.sleep", new_callable=AsyncMock):
-            result = await _async_save_expected(data_with_automation)
-
-        assert mock_store.async_save.call_count == 3
-        assert result is True
-
-    @pytest.mark.asyncio
-    async def test_exponential_backoff_delays(
-        self, data_with_automation: AutomationPauseData, mock_store: MagicMock
-    ) -> None:
-        """Test that retries use exponential backoff timing."""
-        mock_store.async_save = AsyncMock(
-            side_effect=[
-                IOError("Disk full"),
-                IOError("Disk full"),
-                IOError("Disk full"),
-                None,
-            ]
-        )
-
-        with patch("asyncio.sleep", new_callable=AsyncMock) as mock_sleep:
-            await _async_save_expected(data_with_automation)
-
-        expected_delays = [0.1, 0.2, 0.4]
-        actual_delays = [call_args[0][0] for call_args in mock_sleep.call_args_list]
-        assert actual_delays == expected_delays
-
-    @pytest.mark.asyncio
-    async def test_returns_false_after_max_retries(
-        self, data_with_automation: AutomationPauseData, mock_store: MagicMock
-    ) -> None:
-        """Test that save returns False after exhausting retries."""
-        mock_store.async_save = AsyncMock(side_effect=IOError("Persistent failure"))
-
-        with patch("asyncio.sleep", new_callable=AsyncMock):
-            result = await _async_save_expected(data_with_automation)
-
-        assert mock_store.async_save.call_count == 4  # Initial + 3 retries
-        assert result is False
-
-    @pytest.mark.asyncio
-    async def test_does_not_retry_on_non_transient_errors(
-        self, data_with_automation: AutomationPauseData, mock_store: MagicMock
-    ) -> None:
-        """Test that non-transient errors don't trigger retry."""
-        mock_store.async_save = AsyncMock(side_effect=ValueError("Invalid data"))
-
-        with patch("asyncio.sleep", new_callable=AsyncMock) as mock_sleep:
-            result = await _async_save_expected(data_with_automation)
-
-        assert mock_store.async_save.call_count == 1
-        mock_sleep.assert_not_called()
-        assert result is False
-
-    @pytest.mark.asyncio
-    async def test_logs_warning_on_retry(
-        self, data_with_automation: AutomationPauseData, mock_store: MagicMock
-    ) -> None:
-        """Test that retries are logged as warnings."""
-        mock_store.async_save = AsyncMock(side_effect=[IOError("Disk full"), None])
-        mock_logger = MagicMock()
-
-        with patch("asyncio.sleep", new_callable=AsyncMock):
-            await _async_save_expected(data_with_automation, _logger=mock_logger)
-
-        assert mock_logger.warning.called
-        warning_msg = str(mock_logger.warning.call_args)
-        assert "retry" in warning_msg.lower() or "attempt" in warning_msg.lower()
-
-    @pytest.mark.asyncio
-    async def test_logs_error_after_all_retries_exhausted(
-        self, data_with_automation: AutomationPauseData, mock_store: MagicMock
-    ) -> None:
-        """Test that error is logged after all retries fail."""
-        mock_store.async_save = AsyncMock(side_effect=IOError("Persistent failure"))
-        mock_logger = MagicMock()
-
-        with patch("asyncio.sleep", new_callable=AsyncMock):
-            await _async_save_expected(data_with_automation, _logger=mock_logger)
-
-        assert mock_logger.error.called
-
-    @pytest.mark.asyncio
-    async def test_retries_on_oserror(self, data_with_automation: AutomationPauseData, mock_store: MagicMock) -> None:
-        """Test that save retries on OSError."""
-        mock_store.async_save = AsyncMock(side_effect=[OSError("No space left"), None])
-
-        with patch("asyncio.sleep", new_callable=AsyncMock):
-            result = await _async_save_expected(data_with_automation)
-
-        assert mock_store.async_save.call_count == 2
-        assert result is True
-
-
-# =============================================================================
-# VALIDATE ON LOAD TESTS - Source Code Verification
-# =============================================================================
-
-
-class TestValidateOnLoadSourceCode:
-    """Tests that verify the source code has validation logic implemented."""
-
-    def test_has_validation_function(self) -> None:
-        """Test that a validate_stored_data function exists."""
-        source = get_source_code()
-
-        has_validate_func = "validate_stored_data" in source or "validate" in source.lower()
-        assert has_validate_func, "FEATURE NOT IMPLEMENTED: Should have a validate_stored_data function"
-
-    def test_load_stored_validates_schema(self) -> None:
-        """Test that async_load_stored validates the data schema."""
-        source = get_source_code()
-        func = get_async_load_stored_function(source)
-
-        # Should either check isinstance directly or call a validation function
-        checks_type = "isinstance" in func or "validate_stored_data" in func or "validate" in func.lower()
-        # Also verify validation function exists with isinstance checks
-        if "validate_stored_data" in func:
-            checks_type = checks_type and "isinstance" in source
-
-        assert checks_type, "FEATURE NOT IMPLEMENTED: async_load_stored should validate data types with isinstance"
-
-    def test_load_stored_validates_entity_id(self) -> None:
-        """Test that async_load_stored validates entity IDs."""
-        source = get_source_code()
-        func = get_async_load_stored_function(source)
-
-        # Should check entity_id format directly or via validation function
-        checks_entity = (
-            'startswith("automation.' in func or "validate_stored_data" in func or "validate" in func.lower()
-        )
-        # Also verify validation function exists with entity_id check
-        if "validate_stored_data" in func:
-            checks_entity = checks_entity and 'startswith("automation.' in source
-
-        assert checks_entity, "FEATURE NOT IMPLEMENTED: async_load_stored should validate entity IDs are automations"
-
-    def test_load_stored_logs_validation_errors(self) -> None:
-        """Test that validation errors are logged."""
-        source = get_source_code()
-        func = get_async_load_stored_function(source)
-
-        # Should log warnings for invalid entries
-        # Current implementation logs for KeyError/ValueError but should be more comprehensive
-        logs_issues = "_LOGGER.warning" in func or "_LOGGER.error" in func
-        assert logs_issues, "Validation should log warnings for invalid entries"
-
-
-# =============================================================================
-# VALIDATE ON LOAD TESTS - Behavior Specification
-# =============================================================================
-
-
-class TestValidateOnLoadBehavior:
-    """Tests that specify the expected validation behavior."""
-
-    def test_validates_paused_schema_structure(self) -> None:
-        """Test that paused data must be a dict."""
-        mock_logger = MagicMock()
-
-        result = _validate_stored_data(
-            {"paused": ["not", "a", "dict"], "scheduled": {}},
-            _logger=mock_logger,
-        )
-
-        assert result["paused"] == {}
-        assert mock_logger.warning.called
-
-    def test_validates_scheduled_schema_structure(self) -> None:
-        """Test that scheduled data must be a dict."""
-        mock_logger = MagicMock()
-
-        result = _validate_stored_data(
-            {"paused": {}, "scheduled": "not a dict"},
-            _logger=mock_logger,
-        )
-
-        assert result["scheduled"] == {}
-        assert mock_logger.warning.called
-
-    def test_skips_entry_with_invalid_datetime_format(self) -> None:
-        """Test that entries with invalid datetime values are skipped."""
-        now = datetime.now(UTC)
-        mock_logger = MagicMock()
-
-        result = _validate_stored_data(
-            {
-                "paused": {
-                    "automation.valid": {
-                        "friendly_name": "Valid",
-                        "resume_at": (now + timedelta(hours=1)).isoformat(),
-                        "paused_at": now.isoformat(),
-                    },
-                    "automation.invalid": {
-                        "friendly_name": "Invalid",
-                        "resume_at": "not-a-datetime",
-                        "paused_at": now.isoformat(),
-                    },
-                },
-                "scheduled": {},
-            },
-            _logger=mock_logger,
-        )
-
-        assert "automation.valid" in result["paused"]
-        assert "automation.invalid" not in result["paused"]
-
-    def test_skips_entry_with_missing_required_fields(self) -> None:
-        """Test that entries missing required fields are skipped."""
-        now = datetime.now(UTC)
-        mock_logger = MagicMock()
-
-        result = _validate_stored_data(
-            {
-                "paused": {
-                    "automation.valid": {
-                        "friendly_name": "Valid",
-                        "resume_at": (now + timedelta(hours=1)).isoformat(),
-                        "paused_at": now.isoformat(),
-                    },
-                    "automation.missing_resume": {
-                        "friendly_name": "Missing Resume",
-                        "paused_at": now.isoformat(),
-                    },
-                },
-                "scheduled": {},
-            },
-            _logger=mock_logger,
-        )
-
-        assert "automation.valid" in result["paused"]
-        assert "automation.missing_resume" not in result["paused"]
-
-    def test_validates_entity_id_format(self) -> None:
-        """Test that entity IDs are validated to be automation entities."""
-        now = datetime.now(UTC)
-        mock_logger = MagicMock()
-
-        result = _validate_stored_data(
-            {
-                "paused": {
-                    "automation.valid": {
-                        "friendly_name": "Valid",
-                        "resume_at": (now + timedelta(hours=1)).isoformat(),
-                        "paused_at": now.isoformat(),
-                    },
-                    "light.not_automation": {
-                        "friendly_name": "Not an automation",
-                        "resume_at": (now + timedelta(hours=1)).isoformat(),
-                        "paused_at": now.isoformat(),
-                    },
-                },
-                "scheduled": {},
-            },
-            _logger=mock_logger,
-        )
-
-        assert "automation.valid" in result["paused"]
-        assert "light.not_automation" not in result["paused"]
-
-    def test_handles_completely_corrupted_storage(self) -> None:
-        """Test that completely corrupted storage is handled gracefully."""
-        mock_logger = MagicMock()
-
-        result = _validate_stored_data("not even a dict", _logger=mock_logger)
-
-        assert result == {"paused": {}, "scheduled": {}}
-        assert mock_logger.error.called
-
-    def test_validates_numeric_fields_are_non_negative(self) -> None:
-        """Test that numeric fields are validated."""
-        now = datetime.now(UTC)
-        mock_logger = MagicMock()
-
-        result = _validate_stored_data(
-            {
-                "paused": {
-                    "automation.valid": {
-                        "friendly_name": "Valid",
-                        "resume_at": (now + timedelta(hours=1)).isoformat(),
-                        "paused_at": now.isoformat(),
-                        "days": 0,
-                        "hours": 1,
-                        "minutes": 0,
-                    },
-                    "automation.negative": {
-                        "friendly_name": "Negative values",
-                        "resume_at": (now + timedelta(hours=1)).isoformat(),
-                        "paused_at": now.isoformat(),
-                        "days": -1,
-                        "hours": 1,
-                        "minutes": 0,
-                    },
-                },
-                "scheduled": {},
-            },
-            _logger=mock_logger,
-        )
-
-        assert "automation.valid" in result["paused"]
-        assert "automation.negative" not in result["paused"]
-
-    def test_validates_scheduled_disable_at_before_resume_at(self) -> None:
-        """Test that scheduled snooze has disable_at before resume_at."""
-        now = datetime.now(UTC)
-        mock_logger = MagicMock()
-
-        result = _validate_stored_data(
-            {
-                "paused": {},
-                "scheduled": {
-                    "automation.valid": {
-                        "friendly_name": "Valid",
-                        "disable_at": (now + timedelta(hours=1)).isoformat(),
-                        "resume_at": (now + timedelta(hours=2)).isoformat(),
-                    },
-                    "automation.invalid_order": {
-                        "friendly_name": "Invalid Order",
-                        "disable_at": (now + timedelta(hours=2)).isoformat(),
-                        "resume_at": (now + timedelta(hours=1)).isoformat(),
-                    },
-                },
-            },
-            _logger=mock_logger,
-        )
-
-        assert "automation.valid" in result["scheduled"]
-        assert "automation.invalid_order" not in result["scheduled"]
-
-    def test_logs_summary_of_validation_issues(self) -> None:
-        """Test that a summary is logged when validation issues are found."""
-        now = datetime.now(UTC)
-        mock_logger = MagicMock()
-
-        _validate_stored_data(
-            {
-                "paused": {
-                    "automation.invalid1": {"friendly_name": "Invalid 1"},
-                    "automation.invalid2": {
-                        "friendly_name": "Invalid 2",
-                        "resume_at": "bad-date",
-                        "paused_at": now.isoformat(),
-                    },
-                },
-                "scheduled": {},
-            },
-            _logger=mock_logger,
-        )
-
-        # Should log info about number of skipped entries
-        assert mock_logger.info.called
-        info_msg = str(mock_logger.info.call_args)
-        assert "skip" in info_msg.lower() or "invalid" in info_msg.lower()
-
-
-# =============================================================================
-# INTEGRATION TESTS - Both features together
-# =============================================================================
-
-
-class TestPersistenceRobustnessIntegration:
-    """Integration tests for persistence robustness features."""
-
-    @pytest.mark.asyncio
-    async def test_save_failure_does_not_corrupt_runtime_state(self) -> None:
-        """Test that failed saves don't affect runtime data."""
-        mock_store = MagicMock()
-        mock_store.async_save = AsyncMock(side_effect=IOError("Disk error"))
-
-        data = AutomationPauseData(store=mock_store)
-        now = datetime.now(UTC)
-        data.paused["automation.test"] = PausedAutomation(
-            entity_id="automation.test",
-            friendly_name="Test",
-            resume_at=now + timedelta(hours=1),
-            paused_at=now,
-        )
-
-        with patch("asyncio.sleep", new_callable=AsyncMock):
-            await _async_save_expected(data)
-
-        # Runtime data should be unchanged even after save failure
-        assert "automation.test" in data.paused
-
-    def test_validation_allows_partial_recovery(self) -> None:
-        """Test that valid entries are loaded even when some are invalid."""
-        now = datetime.now(UTC)
-        mock_logger = MagicMock()
-
-        result = _validate_stored_data(
-            {
-                "paused": {
-                    "automation.valid1": {
-                        "friendly_name": "Valid 1",
-                        "resume_at": (now + timedelta(hours=1)).isoformat(),
-                        "paused_at": now.isoformat(),
-                    },
-                    "automation.invalid": {
-                        "friendly_name": "Invalid",
-                    },
-                    "automation.valid2": {
-                        "friendly_name": "Valid 2",
-                        "resume_at": (now + timedelta(hours=2)).isoformat(),
-                        "paused_at": now.isoformat(),
-                    },
-                },
-                "scheduled": {},
-            },
-            _logger=mock_logger,
-        )
-
-        # Both valid entries should be loaded
-        assert "automation.valid1" in result["paused"]
-        assert "automation.valid2" in result["paused"]
-        # Invalid entry should be skipped
-        assert "automation.invalid" not in result["paused"]
-
-
-# =============================================================================
-# DELETED AUTOMATION CLEANUP TESTS
-# =============================================================================
-
-
-class TestDeletedAutomationCleanupSourceCode:
-    """Tests that verify deleted automation cleanup is implemented."""
-
-    def test_load_stored_checks_entity_existence_for_paused(self) -> None:
-        """Test that _async_load_stored checks if paused automations still exist."""
-        source = get_source_code()
-        func = get_async_load_stored_function(source)
-
-        # Should check entity existence before loading paused automations
-        # Look for states.get check in the paused automation loading section
-        checks_existence = "states.get" in func or "hass.states.get" in func
-        has_cleanup_log = "Cleaning up deleted automation" in func
-
-        assert checks_existence and has_cleanup_log, (
-            "FEATURE NOT IMPLEMENTED: _async_load_stored should check if paused "
-            "automations still exist and clean up deleted ones"
-        )
-
-    def test_load_stored_checks_entity_existence_for_scheduled(self) -> None:
-        """Test that _async_load_stored checks if scheduled automations still exist."""
-        source = get_source_code()
-        func = get_async_load_stored_function(source)
-
-        # Should check entity existence before loading scheduled snoozes
-        # Look for cleaning up from scheduled storage
-        has_scheduled_cleanup = "Cleaning up deleted automation from scheduled" in func
-
-        assert has_scheduled_cleanup, (
-            "FEATURE NOT IMPLEMENTED: _async_load_stored should check if scheduled "
-            "automations still exist and clean up deleted ones"
-        )
-
-
-class TestDeletedAutomationCleanupBehavior:
-    """Tests that specify the expected cleanup behavior for deleted automations."""
+class TestDeletedAutomationCleanup:
+    """Tests that verify deleted automations are cleaned up from storage on load."""
 
     @pytest.fixture
     def mock_store(self) -> MagicMock:
@@ -986,7 +42,6 @@ class TestDeletedAutomationCleanupBehavior:
         now = datetime.now(UTC)
         data = AutomationPauseData(store=mock_store)
 
-        # Simulate stored data with an automation that no longer exists
         stored_data = {
             "paused": {
                 "automation.deleted": {
@@ -1058,7 +113,6 @@ class TestDeletedAutomationCleanupBehavior:
         now = datetime.now(UTC)
         data = AutomationPauseData(store=mock_store)
 
-        # Simulate stored data with a scheduled snooze for a deleted automation
         stored_data = {
             "paused": {},
             "scheduled": {
@@ -1175,51 +229,132 @@ class TestDeletedAutomationCleanupBehavior:
         assert "automation.existing2" in data.scheduled
 
 
-# =============================================================================
-# SUMMARY: What needs to be implemented
-# =============================================================================
+class TestStorageValidation:
+    """Tests for storage data validation during load."""
 
+    @pytest.fixture
+    def mock_store(self) -> MagicMock:
+        """Create mock store."""
+        store = MagicMock()
+        store.async_save = AsyncMock()
+        return store
 
-class TestImplementationRequirements:
-    """Summary tests that document what needs to be implemented.
+    @pytest.fixture
+    def mock_hass(self) -> MagicMock:
+        """Create mock Home Assistant instance."""
+        hass = MagicMock()
+        hass.services.async_call = AsyncMock()
+        return hass
 
-    These tests will all fail until the features are properly implemented
-    in the source code.
-    """
+    @pytest.mark.asyncio
+    async def test_handles_empty_storage(self, mock_hass: MagicMock, mock_store: MagicMock) -> None:
+        """Test that empty storage is handled gracefully."""
+        from custom_components.autosnooze.coordinator import async_load_stored as _async_load_stored
+        from custom_components.autosnooze.models import AutomationPauseData
 
-    def test_retry_save_implementation_checklist(self) -> None:
-        """Checklist of retry save requirements."""
-        source = get_source_code()
-        const_source = get_const_source_code()
-        func = get_async_save_function(source)
+        data = AutomationPauseData(store=mock_store)
+        mock_store.async_load = AsyncMock(return_value=None)
 
-        # Check if using TRANSIENT_ERRORS constant (which contains IOError/OSError in const.py)
-        uses_transient_const = "TRANSIENT_ERRORS" in func and "IOError" in const_source and "OSError" in const_source
+        await _async_load_stored(mock_hass, data)
 
-        requirements = {
-            "Return bool (True/False)": "return True" in func and "return False" in func,
-            "Retry loop": any(x in func for x in ["for attempt", "for _ in range", "while"]),
-            "Exponential backoff": "asyncio.sleep" in func,
-            "Catch IOError/OSError": "IOError" in func or "OSError" in func or uses_transient_const,
-            "Log warnings on retry": "_LOGGER.warning" in func and "retry" in func.lower(),
-            "Log error after exhaustion": "_LOGGER.error" in func,
+        assert len(data.paused) == 0
+        assert len(data.scheduled) == 0
+
+    @pytest.mark.asyncio
+    async def test_handles_missing_paused_key(self, mock_hass: MagicMock, mock_store: MagicMock) -> None:
+        """Test that missing 'paused' key is handled gracefully."""
+        from custom_components.autosnooze.coordinator import async_load_stored as _async_load_stored
+        from custom_components.autosnooze.models import AutomationPauseData
+
+        data = AutomationPauseData(store=mock_store)
+        mock_store.async_load = AsyncMock(return_value={"scheduled": {}})
+
+        await _async_load_stored(mock_hass, data)
+
+        assert len(data.paused) == 0
+
+    @pytest.mark.asyncio
+    async def test_handles_missing_scheduled_key(self, mock_hass: MagicMock, mock_store: MagicMock) -> None:
+        """Test that missing 'scheduled' key is handled gracefully."""
+        from custom_components.autosnooze.coordinator import async_load_stored as _async_load_stored
+        from custom_components.autosnooze.models import AutomationPauseData
+
+        data = AutomationPauseData(store=mock_store)
+        mock_store.async_load = AsyncMock(return_value={"paused": {}})
+
+        await _async_load_stored(mock_hass, data)
+
+        assert len(data.scheduled) == 0
+
+    @pytest.mark.asyncio
+    async def test_skips_entries_with_invalid_datetime(self, mock_hass: MagicMock, mock_store: MagicMock) -> None:
+        """Test that entries with invalid datetime are skipped."""
+        from custom_components.autosnooze.coordinator import async_load_stored as _async_load_stored
+        from custom_components.autosnooze.models import AutomationPauseData
+
+        now = datetime.now(UTC)
+        data = AutomationPauseData(store=mock_store)
+
+        stored_data = {
+            "paused": {
+                "automation.valid": {
+                    "friendly_name": "Valid",
+                    "resume_at": (now + timedelta(hours=1)).isoformat(),
+                    "paused_at": now.isoformat(),
+                },
+                "automation.invalid": {
+                    "friendly_name": "Invalid",
+                    "resume_at": "not-a-datetime",
+                    "paused_at": now.isoformat(),
+                },
+            },
+            "scheduled": {},
         }
+        mock_store.async_load = AsyncMock(return_value=stored_data)
 
-        missing = [req for req, met in requirements.items() if not met]
+        # Both entities exist
+        mock_state = MagicMock()
+        mock_state.attributes = {"friendly_name": "Test"}
+        mock_hass.states.get.return_value = mock_state
 
-        assert not missing, f"FEATURES NOT IMPLEMENTED in async_save:\n" + "\n".join(f"  - {req}" for req in missing)
+        with patch("custom_components.autosnooze.coordinator.schedule_resume"):
+            await _async_load_stored(mock_hass, data)
 
-    def test_validate_on_load_implementation_checklist(self) -> None:
-        """Checklist of validate on load requirements."""
-        source = get_source_code()
+        # Valid entry should be loaded, invalid should be skipped
+        assert "automation.valid" in data.paused
+        assert "automation.invalid" not in data.paused
 
-        requirements = {
-            "validate_stored_data function exists": "validate_stored_data" in source,
-            "Schema validation (isinstance checks)": "isinstance(paused" in source or "isinstance(stored" in source,
-            "Entity ID validation": 'startswith("automation.' in source,
-            "Logs validation warnings": True,  # Already logs some warnings
-        }
 
-        missing = [req for req, met in requirements.items() if not met]
+class TestSaveFailureRecovery:
+    """Tests for save failure handling."""
 
-        assert not missing, f"FEATURES NOT IMPLEMENTED for validation:\n" + "\n".join(f"  - {req}" for req in missing)
+    @pytest.fixture
+    def mock_store(self) -> MagicMock:
+        """Create mock store."""
+        store = MagicMock()
+        store.async_save = AsyncMock()
+        return store
+
+    @pytest.mark.asyncio
+    async def test_save_failure_does_not_corrupt_runtime_state(self, mock_store: MagicMock) -> None:
+        """Test that failed saves don't affect runtime data."""
+        from custom_components.autosnooze.coordinator import async_save as _async_save
+        from custom_components.autosnooze.models import AutomationPauseData, PausedAutomation
+
+        mock_store.async_save = AsyncMock(side_effect=IOError("Disk error"))
+
+        data = AutomationPauseData(store=mock_store)
+        now = datetime.now(UTC)
+        data.paused["automation.test"] = PausedAutomation(
+            entity_id="automation.test",
+            friendly_name="Test",
+            resume_at=now + timedelta(hours=1),
+            paused_at=now,
+        )
+
+        # Save should fail but not raise
+        await _async_save(data)
+
+        # Runtime data should be unchanged even after save failure
+        assert "automation.test" in data.paused
+        assert data.paused["automation.test"].friendly_name == "Test"
